REF: machine_learning/ > src/: renamed neural network's directory from `machine_learning` to `src`.
