+++ application/app.py
@@ -13,7 +13,7 @@ from dash import html
 import dash.dependencies as dep
 
 from application.components import AppEstuary, EstuaryType, input_check
-from machine_learning.neural_network import NeuralNetwork
+from src.neural_network import NeuralNetwork
 
 # initiate dash-app
 APP = dash.Dash(__name__)
